==> pandas_lesson/02.ipynb <==
#### Day 2 Week 1

import pandas as pd

df = pd.read_csv('crypto_market_data.csv')
df.head()

#### Select closing prices columns

closing_prices = df[['close_btc', 'close_eth', 'close_bnb']]
closing_prices

first_4_days = df.iloc[1:5]
first_4_days

#### Filter high volume days for btc

btc_volume_treshold = df['volume_btc'].quantile(0.75)
high_volume_btc = df[df['volume_btc'] > btc_volume_treshold]
high_volume_btc

btc_threshold = df['volume_btc'].quantile(0.9)

# Complex filtering conditionally with columns selection
high_btc_volume = df.loc[(df['volume_btc'] > btc_threshold) & (df['volume_btc'].notna()), ['volume_btc', 'close_btc']]

# In this case simple filtering -> without columns selection
# high_btc_volume = df[df['volume_btc'] > btc_treshold]

high_btc_volume

#### . Select `volume_btc` and `volume_eth` columns, print the first 5 rows.

volume_btc_eth = df[["volume_btc", "volume_eth"]]
first_5_rows = volume_btc_eth.head().reset_index(drop=True)
volume_str = f"BTC: {volume_btc_eth['volume_btc']}\nETH: {volume_btc_eth['volume_eth']}"
print(volume_str)

#### Use `loc` to select rows where `close_btc` > its median (handle `NaN` with `notna()`).

median_close = df['close_btc'].median()
filtered_df = df.loc[(df['close_btc'].notna()) & (df['close_btc'] > median_close)]
filtered_df

#### Filter days where `close_btc` increased from the previous day (use `diff()`).

filtered_days = df[(df['close_btc'].diff() > 0) & (df['close_btc'].notna()) ]
filtered_days

#### **Mini-Project**: Create a DataFrame with days where `volume_btc` > 90th percentile, save to `high_volume_btc.csv`.

volume_btc_treshould = df['volume_btc'].quantile(0.9)
high_btc_volume = df.loc[df['volume_btc'] > volume_btc_treshould]
df.to_csv("high_btc_volume.csv")
print('Saved to csv file')
high_btc_volume
