==> pandas_lesson/03.ipynb <==
import pandas as pd

df = pd.read_csv('crypto_market_data.csv')
df.head()

#### Day 3 Detect and handle missing values (`NaN`) in `close_btc`

#### Detect missing values

df.isna().sum()

#### Fill with mean

df_filled = df.copy()
df_filled.head()

df_filled['df_filled_mean'] = df_filled['close_btc'].fillna(df['close_btc'].mean())
df_filled[['close_btc', 'df_filled_mean']].head()

#### Interpolate

df_filled['df_interpolated'] = df_filled['close_btc'].interpolate()
df_filled[['close_btc', 'df_filled_mean', 'df_interpolated']].head()

#### Count missing values in `close_btc` and `close_bnb`.

df_filled[['close_btc', 'close_bnb']].isna().sum()

#### Fill missing `close_btc` values with the median.

# Fill some rows with NAN
df_filled.loc[0:3, ['close_bnb']] = float("nan")

medians = df_filled[['close_btc', 'close_bnb']].median()
# df_filled[['close_btc_na', 'close_bnb_na']] = medians.tolist()
# df_filled[['close_btc', 'close_btc_na', 'close_bnb', 'close_bnb_na']].head()

df_filled['close_btc_na'] = medians['close_btc']
df_filled['close_bnb+na'] = medians['close_bnb']
df_filled[['close_btc', 'close_btc_na', 'close_bnb', 'close_bnb_na']].head()

#### Compare the mean of `close_btc` before and after filling (using `df_filled`)

df_filled[['close_btc', 'close_btc_na']].mean()

#### Mini-Project**: Create a DataFrame with no missing `close_btc` (use interpolation), 
#### compute average `close_btc` for first vs. second half of the dataset.

df_filled['close_btc_interpolated'] = df_filled['close_btc'].interpolate()

total_rows = len(df_filled)

mid_point = total_rows // 2

first_half = df_filled.iloc[:mid_point]
second_half = df_filled.iloc[mid_point:]

dif = first_half['close_btc_interpolated'].mean() > second_half['close_btc_interpolated'].mean()

dif
